pyproject.toml
[tool.poetry]
name = "gighub_backend"
version = "0.1.0"
description = ""
authors = ["Julie Park <juliepark242@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.4"
psycopg2 = "^2.9.3"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.1.0"
django-rest-auth = "^0.9.5"
dj-rest-auth = "^2.2.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
mypy = "^0.950"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
